refactor(plugins): add root prompt templates prototype
