+++ plugins/prompt-templates/index.mjs
@@ -1,0 +1,7 @@
+export { createPluginsMethods } from './methods.mjs';
+export { createPluginsComputed } from './computed.mjs';
+export {
+    readPromptTemplatesFromStorage,
+    persistPromptTemplatesToStorage,
+    clearPromptTemplatesStorage
+} from './storage.mjs';
